requirements.txt
streamlit
pandas
python-pptx
deep-translator
openpyxl
